revert to pyinstaller
