+++ resources/cpp/segmentation/CMakeLists.txt
