+++ resources/cpp/segmentation/main.cpp
