+++ src-tauri/src/main.rs
@@ -5,6 +5,7 @@
 
 mod commands;
 mod data;
+mod helpers;
 mod models;
 mod schema;
 mod tasks;
